+++ test/test_indexeddb.js
@@ -56,7 +56,9 @@ describe('IndexedDB tests', function () {
       .then(function () {
         return bag.get(key);
       })
-      .then(null, function (err) { assert.ok(err); });
+      .then(function (val) {
+        assert.isUndefined(val);
+      });
   });
 
 
@@ -82,7 +84,9 @@ describe('IndexedDB tests', function () {
       .then(function () {
         return bag.get(key);
       })
-      .then(null, function (err) { assert.ok(err); });
+      .then(function (val) {
+        assert.isUndefined(val);
+      });
   });
 
 
@@ -109,8 +113,9 @@ describe('IndexedDB tests', function () {
       setTimeout(function () {
         bag.clear(true, function (err) {
           assert.notOk(err);
-          bag.get(key, function (err) {
-            assert.ok(err);
+          bag.get(key, function (err, data) {
+            assert.notOk(err);
+            assert.isUndefined(data);
             done();
           });
         });